refactor(chats-ui): extract and memo prompt input tools
