+++ platform/frontend/src/app/chat/prompt-input.tsx
@@ -4,23 +4,17 @@ import {
   type ChatSkillMetadata,
   E2eTestId,
   getAcceptedFileTypes,
-  getSupportedFileTypesDescription,
-  type ModelInputModality,
-  type SupportedProvider,
   supportsFileUploads,
 } from "@shared";
 import type { ChatStatus } from "ai";
-import { MoreVerticalIcon, PaperclipIcon, XIcon } from "lucide-react";
 import type { FormEvent, KeyboardEvent } from "react";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { toast } from "sonner";
-import { ModelSelectorLogo } from "@/components/ai-elements/model-selector";
 import {
   PromptInput,
   PromptInputAttachment,
   PromptInputAttachments,
   PromptInputBody,
-  PromptInputButton,
   PromptInputCommand,
   PromptInputCommandEmpty,
   PromptInputCommandGroup,
@@ -32,42 +26,21 @@ import {
   PromptInputSpeechButton,
   PromptInputSubmit,
   PromptInputTextarea,
-  PromptInputTools,
-  usePromptInputAttachments,
   usePromptInputController,
 } from "@/components/ai-elements/prompt-input";
-import { ContextIndicator } from "@/components/chat/context-indicator";
-import { InitialAgentSelector } from "@/components/chat/initial-agent-selector";
-import { LlmProviderApiKeySelector } from "@/components/chat/llm-provider-api-key-selector";
-import {
-  ModelSelector,
-  providerToLogoProvider,
-} from "@/components/chat/model-selector";
 import { PlaywrightInstallInline } from "@/components/chat/playwright-install-dialog";
 import { SensitiveDataConfirmDialog } from "@/components/chat/sensitive-data-confirm-dialog";
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
-import { useHasPermissions } from "@/lib/auth/auth.query";
 import { useChatPlaceholder } from "@/lib/chat/chat-placeholder.hook";
 import { conversationStorageKeys } from "@/lib/chat/chat-utils";
-import type { ModelSource } from "@/lib/chat/use-chat-preferences";
-import { useModelSelectorDisplay } from "@/lib/chat/use-model-selector-display.hook";
 import { useFeature } from "@/lib/config/config.query";
-import { useIsMobile } from "@/lib/hooks/use-mobile";
 import { useOrganization } from "@/lib/organization.query";
 import { scanText } from "@/lib/sensitive-data";
 import { useSkillsPaginated } from "@/lib/skills/skill.query";
 import { cn } from "@/lib/utils";
+import {
+  ChatPromptInputTools,
+  type ChatPromptInputToolsProps,
+} from "./prompt-input-tools";
 import {
   buildSkillCommands,
   parseSkillCommand,
@@ -77,42 +50,18 @@ import {
 const CHAT_ATTACHMENT_MAX_BYTES = 50 * 1024 * 1024;
 const CHAT_ATTACHMENT_MAX_MB = CHAT_ATTACHMENT_MAX_BYTES / (1024 * 1024);
 
-export interface ArchestraPromptInputProps {
+export interface ArchestraPromptInputProps
+  extends Omit<ChatPromptInputToolsProps, "textareaRef"> {
   onSubmit: (
     message: PromptInputMessage,
     e: FormEvent<HTMLFormElement>,
     options?: { skill?: ChatSkillMetadata },
   ) => void;
   status: ChatStatus;
-  selectedModel: string;
-  onModelChange: (model: string) => void;
   // Tools integration props
   agentId: string;
-  /** Optional - if not provided, it's initial chat mode (no conversation yet) */
-  conversationId?: string;
-  // API key selector props
-  currentConversationChatApiKeyId?: string | null;
-  currentProvider?: SupportedProvider;
-  /** Selected API key ID for initial chat mode */
-  initialApiKeyId?: string | null;
-  /** Callback for API key change in initial chat mode (no conversation) */
-  onApiKeyChange?: (apiKeyId: string) => void;
-  /** Callback when user selects an API key with a different provider */
-  onProviderChange?: (provider: SupportedProvider, apiKeyId: string) => void;
   // Ref for autofocus
   textareaRef?: React.RefObject<HTMLTextAreaElement | null>;
-  /** Whether file uploads are allowed (controlled by organization setting) */
-  allowFileUploads?: boolean;
-  /** Whether models are still loading - passed to API key selector */
-  isModelsLoading?: boolean;
-  /** Estimated tokens used in the conversation (for context indicator) */
-  tokensUsed?: number;
-  /** Maximum context length of the selected model (for context indicator) */
-  maxContextLength?: number | null;
-  /** Input modalities supported by the selected model (for file type filtering) */
-  inputModalities?: ModelInputModality[] | null;
-  /** Agent's configured LLM API key ID - passed to LlmProviderApiKeySelector */
-  agentLlmApiKeyId?: string | null;
   /** Disable the submit button (e.g., when Playwright setup overlay is visible) */
   submitDisabled?: boolean;
   /** Disable chat input while context compaction is running */
@@ -121,16 +70,6 @@ export interface ArchestraPromptInputProps {
   onCompactConversation?: () => Promise<void> | void;
   /** Whether Playwright setup overlay is visible (for showing Playwright install dialog) */
   isPlaywrightSetupVisible: boolean;
-  /** Current agent ID for agent selector */
-  selectorAgentId?: string | null;
-  /** Fallback display name when the selected agent is not yet present in the cached agent list */
-  selectorAgentName?: string;
-  /** Callback when agent changes */
-  onAgentChange?: (agentId: string) => void;
-  /** Source of the currently selected model (agent, organization, user, or null) */
-  modelSource?: ModelSource | null;
-  /** Callback to reset user model override back to agent/org default */
-  onResetModelOverride?: () => void;
 }
 
 type SlashCommand = {
@@ -182,43 +121,16 @@ const PromptInputContent = ({
   const internalTextareaRef = useRef<HTMLTextAreaElement>(null);
   const textareaRef = externalTextareaRef ?? internalTextareaRef;
   const controller = usePromptInputController();
-  const attachments = usePromptInputAttachments();
   const commandItemRefs = useRef<Array<HTMLDivElement | null>>([]);
   const [activeCommandIndex, setActiveCommandIndex] = useState(0);
   const [dismissedSlashCommandValue, setDismissedSlashCommandValue] = useState<
     string | null
   >(null);
 
-  // Collapsed/expanded state for the model selector (defaults to collapsed = provider icon only)
-  const { isCollapsed: showDefaultLogo, expand: expandModelSelector } =
-    useModelSelectorDisplay({ conversationId });
-
-  const logoProvider = currentProvider
-    ? providerToLogoProvider[currentProvider]
-    : null;
-
-  // Label for the model-source badge. A custom model is a "chat override" when
-  // it is scoped to an existing conversation, and a "user override" otherwise
-  // (the new-chat case, where it reflects the user's own default).
-  const modelSourceLabel =
-    modelSource === "agent"
-      ? "agent"
-      : modelSource === "organization"
-        ? "org"
-        : conversationId
-          ? "chat override"
-          : "user override";
-
   // Derive file upload capabilities from model input modalities
-  const modelSupportsFiles = supportsFileUploads(inputModalities);
+  const showFileUploadButton =
+    allowFileUploads && supportsFileUploads(inputModalities);
   const acceptedFileTypes = getAcceptedFileTypes(inputModalities);
-  const supportedTypesDescription =
-    getSupportedFileTypesDescription(inputModalities);
-
-  // Check if user can update agent settings (to show settings link in tooltip)
-  const { data: canUpdateAgentSettings } = useHasPermissions({
-    agentSettings: ["update"],
-  });
 
   // Chat placeholders from organization settings
   const { data: orgData } = useOrganization();
@@ -246,14 +158,6 @@ const PromptInputContent = ({
       conversationId && onCompactConversation ? [COMPACT_COMMAND] : [];
     return [...compact, ...skillCommands];
   }, [conversationId, onCompactConversation, skillCommands]);
-
-  // RBAC: check if user can see agent picker and provider settings in chat
-  const { data: canSeeAgentPicker } = useHasPermissions({
-    chatAgentPicker: ["enable"],
-  });
-  const { data: canSeeProviderSettings } = useHasPermissions({
-    chatProviderSettings: ["enable"],
-  });
 
   const storageKey = conversationId
     ? conversationStorageKeys(conversationId).draft
@@ -298,12 +202,6 @@ const PromptInputContent = ({
     [controller.textInput],
   );
 
-  const isMobile = useIsMobile();
-
-  // Determine if file uploads should be shown
-  // 1. Organization must allow file uploads (allowFileUploads)
-  // 2. Model must support at least one file type (modelSupportsFiles)
-  const showFileUploadButton = allowFileUploads && modelSupportsFiles;
   // The picker stays open while the user is still typing the command token;
   // once a space is entered they have moved on to the prompt body.
   const isSlashCommandOpen =
@@ -537,14 +435,6 @@ const PromptInputContent = ({
     [showFileUploadButton],
   );
 
-  const handleModelSelectorOpenChange = useCallback((open: boolean) => {
-    if (!open) {
-      setTimeout(() => {
-        textareaRef.current?.focus();
-      }, 100);
-    }
-  }, []);
-
   const submitStatus = status === "error" ? "ready" : status;
 
   return (
@@ -637,281 +527,28 @@ const PromptInputContent = ({
           )}
         </PromptInputBody>
         <PromptInputFooter>
-          <PromptInputTools className="gap-0.5">
-            {/* Mobile: vertical three-dots menu for collapsed toolbar items */}
-            {isMobile &&
-              (showDefaultLogo &&
-              logoProvider &&
-              (modelSource === "agent" || modelSource === "organization") ? (
-                <Button
-                  type="button"
-                  variant="ghost"
-                  size="sm"
-                  className="h-8 px-2"
-                  onClick={expandModelSelector}
-                >
-                  <ModelSelectorLogo
-                    provider={logoProvider}
-                    className="size-4"
-                  />
-                </Button>
-              ) : (
-                <Popover>
-                  <PopoverTrigger asChild>
-                    <Button
-                      type="button"
-                      variant="ghost"
-                      size="sm"
-                      className="h-8 px-2"
-                    >
-                      <MoreVerticalIcon className="size-4" />
-                      <span className="sr-only">More options</span>
-                    </Button>
-                  </PopoverTrigger>
-                  <PopoverContent
-                    side="top"
-                    align="start"
-                    className="w-auto p-3"
-                  >
-                    <div className="flex flex-col gap-3">
-                      {canSeeAgentPicker &&
-                        selectorAgentId !== undefined &&
-                        onAgentChange && (
-                          <div>
-                            <p className="text-xs font-medium text-muted-foreground uppercase tracking-wider mb-1">
-                              Agent
-                            </p>
-                            <InitialAgentSelector
-                              currentAgentId={selectorAgentId}
-                              onAgentChange={onAgentChange}
-                            />
-                          </div>
-                        )}
-                      {canSeeProviderSettings && (
-                        <>
-                          {modelSource && (
-                            <div className="flex items-center gap-1.5">
-                              <Badge
-                                variant="secondary"
-                                className="gap-1 bg-slate-200/70 text-slate-600 dark:bg-slate-700/50 dark:text-slate-300 px-3 py-1 text-xs font-medium"
-                              >
-                                {modelSourceLabel}
-                                {modelSource === "user" &&
-                                  onResetModelOverride && (
-                                    <button
-                                      type="button"
-                                      onClick={onResetModelOverride}
-                                      className="text-muted-foreground hover:text-foreground transition-colors"
-                                      title="Reset to default"
-                                    >
-                                      <XIcon className="size-3" />
-                                    </button>
-                                  )}
-                              </Badge>
-                            </div>
-                          )}
-                          {(conversationId || onApiKeyChange) && (
-                            <div>
-                              <p className="text-xs font-medium text-muted-foreground uppercase tracking-wider mb-1">
-                                Provider API Key
-                              </p>
-                              <LlmProviderApiKeySelector
-                                conversationId={conversationId}
-                                currentProvider={currentProvider}
-                                currentConversationChatApiKeyId={
-                                  conversationId
-                                    ? (currentConversationChatApiKeyId ?? null)
-                                    : (initialApiKeyId ?? null)
-                                }
-                                onApiKeyChange={onApiKeyChange}
-                                onProviderChange={onProviderChange}
-                                isModelsLoading={isModelsLoading}
-                                agentLlmApiKeyId={agentLlmApiKeyId}
-                              />
-                            </div>
-                          )}
-                          <div>
-                            <p className="text-xs font-medium text-muted-foreground uppercase tracking-wider mb-1">
-                              Model
-                            </p>
-                            <ModelSelector
-                              selectedModel={selectedModel}
-                              onModelChange={onModelChange}
-                              apiKeyId={
-                                conversationId
-                                  ? currentConversationChatApiKeyId
-                                  : initialApiKeyId
-                              }
-                            />
-                          </div>
-                        </>
-                      )}
-                      {tokensUsed > 0 && maxContextLength && (
-                        <div>
-                          <p className="text-xs font-medium text-muted-foreground uppercase tracking-wider mb-1">
-                            Context
-                          </p>
-                          <ContextIndicator
-                            tokensUsed={tokensUsed}
-                            maxTokens={maxContextLength}
-                            size="sm"
-                          />
-                        </div>
-                      )}
-                    </div>
-                  </PopoverContent>
-                </Popover>
-              ))}
-
-            {/* File attachment button - always visible */}
-            {showFileUploadButton ? (
-              <Tooltip>
-                <TooltipTrigger asChild>
-                  <Button
-                    type="button"
-                    variant="ghost"
-                    size="sm"
-                    className="h-8 px-2"
-                    onClick={() => attachments.openFileDialog()}
-                    data-testid={E2eTestId.ChatFileUploadButton}
-                  >
-                    <PaperclipIcon className="size-4" />
-                    <span className="sr-only">Attach files</span>
-                  </Button>
-                </TooltipTrigger>
-                {supportedTypesDescription && (
-                  <TooltipContent side="top" sideOffset={4}>
-                    Supports: {supportedTypesDescription}
-                  </TooltipContent>
-                )}
-              </Tooltip>
-            ) : (
-              <Tooltip>
-                <TooltipTrigger asChild>
-                  <span
-                    className="inline-flex cursor-pointer"
-                    data-testid={E2eTestId.ChatDisabledFileUploadButton}
-                  >
-                    <PromptInputButton disabled>
-                      <PaperclipIcon className="size-4" />
-                    </PromptInputButton>
-                  </span>
-                </TooltipTrigger>
-                <TooltipContent side="top" sideOffset={4}>
-                  {!allowFileUploads ? (
-                    canUpdateAgentSettings ? (
-                      <span>
-                        File uploads are disabled.{" "}
-                        <a
-                          href="/settings/agents"
-                          className="underline hover:no-underline"
-                          aria-label="Enable file uploads in agent settings"
-                        >
-                          Enable in settings
-                        </a>
-                      </span>
-                    ) : (
-                      "File uploads are disabled by your administrator"
-                    )
-                  ) : (
-                    "This model does not support file uploads"
-                  )}
-                </TooltipContent>
-              </Tooltip>
-            )}
-
-            {/* Desktop: inline toolbar items */}
-            {!isMobile && (
-              <>
-                {canSeeAgentPicker &&
-                  selectorAgentId !== undefined &&
-                  onAgentChange && (
-                    <InitialAgentSelector
-                      currentAgentId={selectorAgentId}
-                      currentAgentName={selectorAgentName}
-                      onAgentChange={onAgentChange}
-                    />
-                  )}
-                {!canSeeProviderSettings ? null : showDefaultLogo &&
-                  logoProvider &&
-                  (modelSource === "agent" ||
-                    modelSource === "organization") ? (
-                  <Button
-                    type="button"
-                    variant="ghost"
-                    size="sm"
-                    className="h-8 px-2"
-                    onClick={expandModelSelector}
-                  >
-                    <ModelSelectorLogo
-                      provider={logoProvider}
-                      className="size-4"
-                    />
-                  </Button>
-                ) : (
-                  <div className="flex items-center h-8 rounded-full border border-border bg-muted/50 overflow-hidden">
-                    {(conversationId || onApiKeyChange) && (
-                      <LlmProviderApiKeySelector
-                        conversationId={conversationId}
-                        currentProvider={currentProvider}
-                        currentConversationChatApiKeyId={
-                          conversationId
-                            ? (currentConversationChatApiKeyId ?? null)
-                            : (initialApiKeyId ?? null)
-                        }
-                        onApiKeyChange={onApiKeyChange}
-                        onProviderChange={onProviderChange}
-                        isModelsLoading={isModelsLoading}
-                        agentLlmApiKeyId={agentLlmApiKeyId}
-                        onOpenChange={(open) => {
-                          if (!open) {
-                            setTimeout(() => {
-                              textareaRef.current?.focus();
-                            }, 100);
-                          }
-                        }}
-                      />
-                    )}
-                    <ModelSelector
-                      selectedModel={selectedModel}
-                      onModelChange={onModelChange}
-                      onOpenChange={handleModelSelectorOpenChange}
-                      apiKeyId={
-                        conversationId
-                          ? currentConversationChatApiKeyId
-                          : initialApiKeyId
-                      }
-                    />
-                    {modelSource && (
-                      <Badge
-                        variant="secondary"
-                        className="ml-1 mr-2 gap-1 bg-slate-200/70 text-slate-600 dark:bg-slate-700/50 dark:text-slate-300 px-3 py-1 text-xs font-medium"
-                      >
-                        {modelSourceLabel}
-                        {modelSource === "user" && onResetModelOverride && (
-                          <button
-                            type="button"
-                            onClick={onResetModelOverride}
-                            className="text-muted-foreground hover:text-foreground transition-colors"
-                            title="Reset to default"
-                          >
-                            <XIcon className="size-3" />
-                          </button>
-                        )}
-                      </Badge>
-                    )}
-                  </div>
-                )}
-                {tokensUsed > 0 && maxContextLength && (
-                  <ContextIndicator
-                    tokensUsed={tokensUsed}
-                    maxTokens={maxContextLength}
-                    size="sm"
-                  />
-                )}
-              </>
-            )}
-          </PromptInputTools>
+          <ChatPromptInputTools
+            selectedModel={selectedModel}
+            onModelChange={onModelChange}
+            conversationId={conversationId}
+            currentConversationChatApiKeyId={currentConversationChatApiKeyId}
+            currentProvider={currentProvider}
+            initialApiKeyId={initialApiKeyId}
+            onApiKeyChange={onApiKeyChange}
+            onProviderChange={onProviderChange}
+            allowFileUploads={allowFileUploads}
+            isModelsLoading={isModelsLoading}
+            tokensUsed={tokensUsed}
+            maxContextLength={maxContextLength}
+            inputModalities={inputModalities}
+            agentLlmApiKeyId={agentLlmApiKeyId}
+            selectorAgentId={selectorAgentId}
+            selectorAgentName={selectorAgentName}
+            onAgentChange={onAgentChange}
+            modelSource={modelSource}
+            onResetModelOverride={onResetModelOverride}
+            textareaRef={textareaRef}
+          />
           <div className="flex items-center gap-2">
             <PromptInputSpeechButton
               textareaRef={textareaRef}
